Bugfix and improvements     Features:     - added cxxCall for Environment     - added conversion for std::string in Vcast     Bugfix:     - added required default branches to switch statements     - added noexcept to various constructors/functions/operators     - added missing virtual destructors     - added missing override specifiers     - fixed use after free     - fixed memory leaks     - fixed logic mistakes     - fixed some opcodes not generate exception handling (handle_catch_2 and handle_catch_3)     - now CaptureStackTrace___ not throw segmentation exception when access to unsafe external stack(read)     - fixed arguments order for proxy functions

+++ src/run_time/tasks/classes/task/task_def.cpp
@@ -277,7 +277,7 @@ namespace art {
             try_count++;
             id++;
         }
-        if (id == -1)
+        if (id == (uint16_t)-1)
             goto is_not_id;
         glob.binded_workers[id].allow_implicit_start = allow_implicit_start;
         glob.binded_workers[id].fixed_size = (bool)fixed_count;
